Account and Comments updates

+++ backend/src/main/java/com/example/backend/Account/Account.java
@@ -12,7 +12,7 @@ public class Account {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "firstname")
     private String firstname;
@@ -40,11 +40,11 @@ public class Account {
         this.rewardspoints = rewardspoints;
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
